agent-deploy/pyproject.toml
[project]
name = "agent-deploy"
version = "0.1.0"
description = "Ship any Python agent to production with a single command"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "typer>=0.12.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "sse-starlette>=2.1.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.5.0",
    "structlog>=24.4.0",
    "opentelemetry-api>=1.27.0",
    "opentelemetry-sdk>=1.27.0",
    "opentelemetry-exporter-otlp>=1.27.0",
    "prometheus-client>=0.21.0",
    "docker>=7.1.0",
    "kubernetes>=31.0.0",
    "rich>=13.8.0",
    "watchfiles>=0.24.0",
    "httpx>=0.27.0",
    "pyyaml>=6.0.2",
]

[project.scripts]
adeploy = "agent_deploy.cli.main:app"
agent-deploy = "agent_deploy.cli.main:app"

[tool.uv]
dev-dependencies = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "mypy>=1.11.0",
    "ruff>=0.6.0",
    "testcontainers>=4.8.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
strict = true
python_version = "3.11"
